refactor(frontend): organize feature modules
