--- frontend/src/app/app.routes.ts
@@ -1,24 +1,0 @@
-import { Routes } from '@angular/router';
-import { authGuard } from './services/auth.guard';
-
-export const routes: Routes = [
-  {
-    path: 'login',
-    loadComponent: () => import('./login/login.page').then((m) => m.LoginPage),
-  },
-  {
-    path: 'register',
-    loadComponent: () =>
-      import('./register/register.page').then((m) => m.RegisterPage),
-  },
-  {
-    path: 'users',
-    canActivate: [authGuard],
-    loadComponent: () => import('./users/users.page').then((m) => m.UsersPage),
-  },
-  {
-    path: '',
-    redirectTo: 'login',
-    pathMatch: 'full',
-  },
-];
