+++ frontend/src/app/auth/auth.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { IonicModule } from '@ionic/angular';
+import { ReactiveFormsModule } from '@angular/forms';
+import { AuthRoutingModule } from './auth-routing.module';
+import { LoginPage } from './pages/login/login.page';
+import { RegisterPage } from './pages/register/register.page';
+
+@NgModule({
+  declarations: [LoginPage, RegisterPage],
+  imports: [CommonModule, IonicModule, ReactiveFormsModule, AuthRoutingModule],
+})
+export class AuthModule {}
