+++ frontend/src/app/auth/guards/auth.guard.ts
@@ -1,12 +1,12 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { AuthService } from './auth.service';
+import { AuthService } from '../services/auth.service';
 
 export const authGuard: CanActivateFn = () => {
   const auth = inject(AuthService);
   const router = inject(Router);
   if (!auth.isLoggedIn()) {
-    router.navigateByUrl('/login');
+    router.navigateByUrl('/auth/login');
     return false;
   }
   return true;
