+++ frontend/src/app/auth/pages/register/register.page.scss
