+++ frontend/src/app/core/services/ui.service.ts
@@ -1,8 +1,5 @@
 import { Injectable } from '@angular/core';
-import {
-  AlertController,
-  ToastController,
-} from '@ionic/angular/standalone';
+import { AlertController, ToastController } from '@ionic/angular';
 
 @Injectable({ providedIn: 'root' })
 export class UiService {
