--- frontend/src/app/home/home.page.html
@@ -1,16 +1,0 @@
-<ion-header translucent>
-  <ion-toolbar>
-    <ion-title>Home</ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content fullscreen class="ion-padding">
-  <ion-card class="hero-card ion-text-center">
-    <ion-card-header>
-      <h1>Welcome</h1>
-    </ion-card-header>
-    <ion-card-content>
-      <p>Manage your users easily.</p>
-    </ion-card-content>
-  </ion-card>
-</ion-content>
