+++ frontend/src/app/shared/shared.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { IonicModule } from '@ionic/angular';
+
+@NgModule({
+  exports: [CommonModule, IonicModule],
+})
+export class SharedModule {}
