+++ frontend/src/app/users/components/edit-user-modal/edit-user-modal.component.ts
@@ -1,19 +1,12 @@
 import { Component, EventEmitter, Output } from '@angular/core';
 import { IonicModule } from '@ionic/angular';
 import { CommonModule } from '@angular/common';
-import {
-  ReactiveFormsModule,
-  FormBuilder,
-  FormGroup,
-  Validators,
-} from '@angular/forms';
-import { UserService, User } from '../services/user.service';
-import { UiService } from '../services/ui.service';
+import { ReactiveFormsModule, FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UserService, User } from '../../services/user.service';
+import { UiService } from '../../../core/services/ui.service';
 
 @Component({
   selector: 'app-edit-user-modal',
-  standalone: true,
-  imports: [IonicModule, CommonModule, ReactiveFormsModule],
   templateUrl: './edit-user-modal.component.html',
   styleUrls: ['./edit-user-modal.component.scss'],
 })
